[Manhuagui] Replace locale query to ORIGINAL_LANGUAGE
ORIGINAL_LANGUAGE is literally for the language of the original version
of one work, not LANGUAGE.

After replacement, Kotatsu app may show both language and original
language in the search filter bottom menu, but the former one impacts
nothing to the search result.

Does not pass locale test, that is, does not fix "unsupported for
ORIGINAL_LANGUAGE" or anything similar. Locale test enforces two
LANGUAGE and one ORIGINAL_LANGUAGE criterions **as long as** the parser
has any availabe locale, but not all parsers support both of them simul-
taneouosly, neither of exclusive ones. Due to this kind of disability of
locale test, it will not be fixed temporarily.